backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "openai>=1.65.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "uuid>=1.30",
    "uvicorn>=0.34.0",
]
